lab04/service-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
requests==2.31.0
pymongo==4.5.0
python-dotenv==1.0.0 